client: Don't exit the cliente whens erver disconnects
Now we just try to reconnect to server forever, having a timer of two seconds. The attribute "connected" is
our flag to set when the connection is active. Some changes were needed to achieve this, as changing the cliente ctor,
use server_connect instead of send_register_message(who was killed in this commit) and everything seems to work nicely now.